Modifications to the depth sensor efficiency

+++ Peripherals/Peripheral2/lib/motor_driver/motor_driver.h
@@ -17,7 +17,7 @@ public:
    void init();
    void stepMotor(int thisStep);
    void moveSteps(int steps);
-   void moveLength(uint16_t mm);
+   void moveLength(int mm);
    void tick();
 };
